requirements.txt
Django==4.0.1
django-environ==0.10.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
djangorestframework==3.12.4